apps/vicaran-agent/vicaran_agent/utils/tracing.py
"""
OpenTelemetry tracing utilities for Google Cloud.
"""

from collections.abc import Sequence

from opentelemetry.sdk.trace import ReadableSpan
from opentelemetry.sdk.trace.export import SpanExporter, SpanExportResult


class CloudTraceLoggingSpanExporter(SpanExporter):
    """
    OpenTelemetry span exporter for Google Cloud Logging.
    """

    def __init__(self, project_id: str, service_name: str) -> None:
        self.project_id = project_id
        self.service_name = service_name
        self.tracer = f"projects/{project_id}/traces/{service_name}"

    def export(self, spans: Sequence[ReadableSpan]) -> SpanExportResult:
        # This is a simplified implementation. A production-ready exporter
        # would typically send data to the Cloud Trace API.
        for span in spans:
            print(f"Exporting span: {span.name}")
        return SpanExportResult.SUCCESS

    def shutdown(self) -> None:
        pass
